Enhance GitHub Actions workflow by adding Android NDK setup and Isar build steps
This commit introduces steps to read the Android NDK version from a file and set up the NDK in the workflow. Additionally, it adds a setup step for Isar, including installation of necessary tools and execution of build scripts for Android targets. These changes improve the build process for Android applications using Flutter and Isar.